+++ car-show/src/components/Comment.jsx
@@ -1,17 +1,27 @@
 import { useState, useEffect } from 'react'
 import axios from 'axios'
+import CommentForm from './CommentForm'
 
-const Comment = (reponse) => {
-  const [comments, setComments] = useState([])
+const Comment = ({ carList }) => {
+  // const [comments, setComments] = useState([])
+  // const getAllCommentsForCar = async () => {
+  //   const response = await axios.get(`http://localhost:3001/comment/all/${}`)
+  //   console.log(response.data.content)
+  //   setComments(response.data.content)
+  // }
 
-  const getAllComments = async () => {
-    const response = await axios.get('http://localhost:3001/comment/all')
-    setComments(reponse.data.comments)
-  }
+  // useEffect(() => {
+  //   getAllCommentsForCar()
+  // }, [])
 
   return (
-    <div></div>
-  )
+    <div>
+      <section>
+        Comments:
+      </section>
+    </div>
+
+    )
 }
 
 export default Comment
